Old things delete
Old things delete

+++ eclipse_build_test_libjamesdsp/jni/Biquad.h
@@ -121,12 +121,12 @@ protected:
   void applyScale (double scale);
 
 public:
-  int64_t m_a0;
-  int64_t m_a1;
-  int64_t m_a2;
-  int64_t m_b1;
-  int64_t m_b2;
-  int64_t m_b0;
+  double m_a0;
+  double m_a1;
+  double m_a2;
+  double m_b1;
+  double m_b2;
+  double m_b0;
 };
 
 //------------------------------------------------------------------------------
